requirements.txt
pydantic~=2.11.0
fastapi~=0.115.0
uvicorn~=0.23.0
jinja2~=3.1.6
google-generativeai~=0.8.5
python-dotenv~=1.1.0